Move auth and options helpers to seam.py level
