+++ seam/parse_options.py
@@ -1,7 +1,7 @@
 import os
 from typing import Optional
 
-from seam.utils.auth import get_auth_headers
+from seam.auth import get_auth_headers
 
 DEFAULT_ENDPOINT = "https://connect.getseam.com"
 
@@ -55,3 +55,44 @@ def get_endpoint_from_env():
         )
 
     return seam_endpoint or seam_api_url
+
+
+class SeamHttpInvalidOptionsError(Exception):
+    def __init__(self, message):
+        super().__init__(f"SeamHttp received invalid options: {message}")
+
+
+def is_seam_http_options_with_api_key(
+    api_key: Optional[str] = None,
+    personal_access_token: Optional[str] = None,
+) -> bool:
+    if api_key is None:
+        return False
+
+    if personal_access_token is not None:
+        raise SeamHttpInvalidOptionsError(
+            "The personal_access_token option cannot be used with the api_key option"
+        )
+
+    return True
+
+
+def is_seam_http_options_with_personal_access_token(
+    personal_access_token: Optional[str] = None,
+    api_key: Optional[str] = None,
+    workspace_id: Optional[str] = None,
+) -> bool:
+    if personal_access_token is None:
+        return False
+
+    if api_key is not None:
+        raise SeamHttpInvalidOptionsError(
+            "The api_key option cannot be used with the personal_access_token option"
+        )
+
+    if workspace_id is None:
+        raise SeamHttpInvalidOptionsError(
+            "Must pass a workspace_id when using a personal_access_token"
+        )
+
+    return True
