+++ seam/token.py
